Refactoring Gulpfile using dependency injection
