+++ gulp/tasks/clean.js
@@ -1,0 +1,7 @@
+module.exports = function (gulp, rimraf) {
+
+  gulp.task('clean', function() {
+    return gulp.src('./temp/', { read: false }).pipe(rimraf());
+  });
+
+};
