+++ gulp/tasks/default.js
@@ -1,0 +1,5 @@
+module.exports = function (gulp, runSequence) {
+  gulp.task('default', function(cb){
+    runSequence('clean', ['copy:images', 'copy:md'], 'doc:pdf:all', cb);
+  });
+};
