+++ package.json
@@ -11,11 +11,13 @@
     "url": "git://github.com/mgechev/angularjs-patterns.git"
   },
   "author": "mgechev",
-  "contributors": [{ 
-    "name" : "Wassim Chegham", 
-    "email" : "maneki.nekko@gmail.com", 
-    "url" : "https://github.com/manekinekko/"
-  }],
+  "contributors": [
+    {
+      "name": "Wassim Chegham",
+      "email": "maneki.nekko@gmail.com",
+      "url": "https://github.com/manekinekko/"
+    }
+  ],
   "license": "MIT",
   "gitHead": "2009434a6dfdbe5f06f81253fb853840af753b36",
   "readmeFilename": "README.md",
@@ -28,5 +30,8 @@
     "gulp-rimraf": "^0.1.1",
     "markdown-styles": "https://github.com/mgechev/markdown-styles/tarball/master",
     "run-sequence": "^1.1.1"
+  },
+  "dependencies": {
+    "gulp-di": "0.0.1"
   }
 }
